requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
async-lru==2.0.4
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
build==0.10.0
CacheControl==0.13.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
crashtest==0.4.1
cryptography==41.0.2
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
dnspython==2.3.0
dulwich==0.21.5
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.2
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.7.0
installer==0.7.0
ipykernel==6.23.1
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jaraco.classes==3.2.3
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.24.0
keyring==24.2.0
kiwisolver==1.4.4
langchain==0.0.137
LinearRegression==0.0.1
loguru==0.7.0
Markdown==3.4.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==2.0.5
more-itertools==9.1.0
msgpack==1.0.5
mss==9.0.1
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
outcome==1.2.0
overrides==7.4.0
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pefile==2023.2.7
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pinecone-client==2.2.1
pipenv==2023.8.26
pkginfo==1.9.6
platformdirs==3.10.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
portalocker==2.7.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
PyAutoGUI==0.9.53
pycparser==2.21
pydantic==1.10.7
PyGetWindow==0.0.9
Pygments==2.15.1
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
PyMsgBox==1.0.9
pyparsing==3.0.9
PyPDF2==3.0.1
pyperclip==1.8.2
pyproject_hooks==1.0.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyQt5-stubs==5.15.6.0
PyQt6==6.4.1
PyQt6-Qt6==6.4.2
PyQt6-sip==13.4.1
PyRect==0.2.0
pyrsistent==0.19.3
pyscreenshot==3.1
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python3-xlib==0.15
pythonautogui==0.0.2
pytweening==1.0.4
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
qtwidgets==1.1
rapidfuzz==2.15.1
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
screen-brightness-control==0.18.0
SecretStorage==3.3.3
selenium==4.8.0
Send2Trash==1.8.2
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.48
stack-data==0.6.2
tenacity==8.2.2
terminado==0.17.1
TimerThread==0.1.1
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.8
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
trove-classifiers==2023.5.24
typing-inspect==0.8.0
typing_extensions==4.5.0
uri-template==1.2.0
urllib3==1.26.16
virtualenv==20.24.3
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.7
win32-setctime==1.1.0
winotify==1.1.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.15.0
